requirements.txt
future~=0.18.2
opencv-python~=4.5.1.48
numpy~=1.20.1
pyzbar~=0.1.8
Pillow~=8.2.0
matplotlib~=3.4.3
pandas~=1.3.2
img2pdf~=0.4.1
qrcode~=6.1
openpyxl~=3.0.7
regex~=2021.8.21
pyrebase4~=4.5.0
arabic_reshaper~=2.1.3
python-bidi~=0.4.2